Use __builtin_align_up for PAD() if available
Clang 10+ includes builtins for checking and adjusting alignment without
casting from pointer to integer:
https://clang.llvm.org/docs/LanguageExtensions.html#alignment-builtins

These builtins provide better argument checking (for example they check if
the second argument is a power-of-two) and they are type-preserving.
Additionally, they are defined to work for pointer arguments so we could
remove pointer-to-integer casts that adversely affect the compiler's
analysis passes.